refactor(server): extract /ohlc endpoints into modular REST structure with middlewares
